// src/enums/category.enum.ts
export enum CategoryEnum {
  HOUSING = "housing",
  GROCERIES = "groceries",
  TRANSPORTATION = "transportation",
  HEALTHCARE = "healthcare",
  ENTERTAINMENT = "entertainment",
  SHOPPING = "shopping",
  SPORTS = "sports",
  TRAVEL = "travel",
  DEBT = "debt",
  SAVINGS = "savings",
  EDUCATION = "education",
}